Try to fix another build error

+++ source/blender/draw/intern/draw_manager_profiling.cc
@@ -180,7 +180,7 @@ void DRW_stats_reset()
 
         timer->time_average = timer->time_average * (1.0 - GPU_TIMER_FALLOFF) +
                               time * GPU_TIMER_FALLOFF;
-        timer->time_average = std::min(timer->time_average, 1000000000ul);
+        timer->time_average = std::min(timer->time_average, uint64_t(1000000000));
       }
       else {
         timer->time_average = lvl_time[timer->lvl + 1];
